requirement.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
numpy==1.24.2
oauthlib==3.2.2
Pillow==9.4.0
pycodestyle==2.10.0
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
tomli==2.0.1
tweepy==4.14.0
urllib3==1.26.14
